Camada application e controller

+++ APICredito.Application/AutoMapper/PropostaCreditoProfile.cs
@@ -9,7 +9,7 @@ namespace APICredito.Application.AutoMapper
         public PropostaCreditoProfile() 
         {
             CreateMap<PropostaCreditoViewModel, PropostaCredito>()
-                 .ConstructUsing(pc => new PropostaCredito(pc.ValorCredito, pc.TipoCredito,pc.QuantidadeParcelas,pc.DataPrimeiroVencimento));
+                 .ConstructUsing(pc => new PropostaCredito(pc.ValorCredito, (int)pc.TipoCredito,pc.QuantidadeParcelas,pc.DataPrimeiroVencimento));
 
             CreateMap<PropostaCredito, PropostaCreditoViewModel>();
